requirements.txt
requests
discord.py
asyncio
tldextract
urlextract